requirements.txt
python-dotenv
langextract
